Tools/CrewAI/pyproject.toml
[project]
name = "enlisted-crew"
version = "0.1.0"
description = "CrewAI multi-agent system for Enlisted mod development"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.95.0",
    "crewai-tools>=0.17.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
enlisted-crew = "enlisted_crew.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/enlisted_crew"]

[tool.hatch.build]
include = [
    "src/enlisted_crew/**/*.py",
    "src/enlisted_crew/config/**/*.yaml",
]
